+++ initpot.h
@@ -25,8 +25,8 @@ public:
 protected:
     const double m_m;
     unsigned int m_anglmomentum;
-    int m_spin;
-    InitialPot(double m, unsigned int l, int j);
+    double m_spin;
+    InitialPot(double m, unsigned int l, double j);
 
 };
 
@@ -40,7 +40,7 @@ class WSaxPot: public InitialPot //derived class
 public:
     WSaxPot(double Rn, double a0, double m);
     WSaxPot(double Rn, double a0, double m, unsigned int l);
-    WSaxPot(double Rn, double a0, double m, unsigned int l, int j);
+    WSaxPot(double Rn, double a0, double m, unsigned int l, double j);
     double potential(double x) const override;
     
     std::unique_ptr<InitialPot> clone() const override;
@@ -61,7 +61,7 @@ class HOPot: public InitialPot //derived class
 public:
     HOPot(double m); //l is assumed ground state
     HOPot(double m, unsigned int l);
-    HOPot(double m, unsigned int l, int j);
+    HOPot(double m, unsigned int l, double j);
     double potential(double x) const override;
 
     std::unique_ptr<InitialPot> clone() const override;
@@ -80,7 +80,7 @@ class CoPot: public InitialPot //derived class
 public:
     CoPot(int Z, double m);
     CoPot(int Z, double m, unsigned int l);
-    CoPot(int Z, double m, unsigned int l, int j);
+    CoPot(int Z, double m, unsigned int l, double j);
     double potential(double x) const override;
 
     std::unique_ptr<InitialPot> clone() const override;
@@ -134,7 +134,7 @@ friend TotPot& operator+(const InitialPot& rhsPotential);*/
 class PotOut: public InitialPot
 {
 public:
-    PotOut(const KohnShamInverse& outpot, double m, unsigned int l, int j);
+    PotOut(const KohnShamInverse& outpot, double m, unsigned int l, double j);
 	PotOut(const KohnShamInverse& outpot, double m, unsigned int l);
     PotOut(const KohnShamInverse& outpot, double m); //l is assumed ground state; j is zero
 	double potential (double x) const override;
@@ -159,7 +159,7 @@ class TestPot: public InitialPot
 public:
 	TestPot(double m); //l is assumed ground state
 	TestPot(double m, unsigned int l);
-    TestPot(double m, unsigned int l, int j);
+    TestPot(double m, unsigned int l, double j);
     double potential(double x) const override;
 
     std::unique_ptr<InitialPot> clone() const override;
